fix(android): cast SHUT_RDWR to Int32 for Bionic's shutdown() signature
On Android's Bionic libc, SHUT_RDWR imports as Int rather than Int32,
breaking the build against Glibc/Darwin where it's already Int32.

+++ Sources/SwiftSFTP/Layer 0/Methods/Convenience/SessionHandshakeTCP.swift
@@ -12,7 +12,7 @@ public typealias SwiftSFTPSocket = Int32
 /// - Parameter socket: The socket descriptor to shut down.
 /// - Throws: `POSIXError` when the descriptor could not be shut down.
 public func ShutdownSocket(_ socket: SwiftSFTPSocket) throws {
-    guard shutdown(socket, SHUT_RDWR) == 0 else {
+    guard shutdown(socket, Int32(SHUT_RDWR)) == 0 else {
         if errno == ENOTCONN {
             return
         }
